Move tests for repo.getRepoDetailsFromRemote for test_repo
commit-id:4c6364e1
